app,lib: Fix minor typos and issues
app: Added free-fall disable function

app: Fixed minor typos

app: Fixed typo for STM32CubeIDE

app: Removed unused led/Rx functions

lib,app: Moved functions to more appropriate places

lib: Set accelerometer priority to lowest value

app: Moved ACC handler to own function

app,lib: Fixed confusing comments

app: Readded RX function
