+++ Software/app/freefall_lorawan/.cproject
@@ -52,7 +52,7 @@
 									<listOptionValue builtIn="false" value="../../"/>
 									<listOptionValue builtIn="false" value="../conf"/>
 									<listOptionValue builtIn="false" value="../lib_dependency"/>
-									<listOptionValue builtIn="false" value="../../../target/STM32WLxx_HAL_Driver/CMSIS/core/include"/>
+									<listOptionValue builtIn="false" value="../../../target/STM32WLxx_HAL_Driver/CMSIS/Core/Include"/>
 									<listOptionValue builtIn="false" value="../../../target/STM32WLxx_HAL_Driver/Inc"/>
 									<listOptionValue builtIn="false" value="../../../lib/STM32WLxx_LoRaWAN/LoRaWAN/Crypto"/>
 									<listOptionValue builtIn="false" value="../../../lib/STM32WLxx_LoRaWAN/LoRaWAN/LmHandler"/>
